[FIX] Player now have cooldowns to shoot

+++ ECS/src/components/ShootCooldown.cpp
@@ -2,7 +2,7 @@
 
 namespace Ecs {
 
-    ShootCD::ShootCD(int cd)
+    ShootCD::ShootCD(float cd)
         : _cd(cd)
     { _type = Constants::SHOOTCD; }
     ShootCD::~ShootCD() {}
